Renaming from lmexporter to logicmonitorexporter
