+++ exporter/logicmonitorexporter/config_test.go
@@ -1,4 +1,4 @@
-package lmexporter
+package logicmonitorexporter
 
 import (
 	"testing"
