+++ exporter/logicmonitorexporter/session.go
@@ -1,4 +1,4 @@
-package lmexporter
+package logicmonitorexporter
 
 import (
 	"bytes"
@@ -154,7 +154,6 @@ func (c *LMhttpClient) MakeRequest(version, method, baseURI, uri, configURL stri
 	if pBytes == nil {
 		pBytes = bytes.NewBuffer(nil)
 	}
-	fmt.Println("Payload :", pBytes.String())
 	req, err = http.NewRequest(method, fullURL, pBytes)
 	if err != nil {
 		return nil, fmt.Errorf("creation of request failed with error %w", err)
